requirements.txt
diffusers
accelerate
transformers
insightface
onnxruntime
streamlit
torch
Pillow
-e .